Start using Vue
