+++ resources/index.html
@@ -14,6 +14,7 @@
     </head>
     <body>
         <main class="container">
+            <div id="app"></div>
             <div id="timerDisplay">10:00</div>
             <div class="cycle-settings">
                 <input id="timerValue" type="number" value="600" min="1" />
